backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
bcrypt==3.2.0
pandas
scikit-learn
python-jose[cryptography]
python-multipart
requests
emails
apscheduler
numpy
# Testing
pytest
pytest-asyncio
httpx
pytest-cov
# Migrations
alembic
# AI Assistant
google-generativeai>=0.8.0
